booking_service/requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
drf-yasg==1.21.11
PyJWT==2.10.1
dj-database-url==3.0.1
requests==2.32.5
pika==1.3.2
django-cors-headers==4.9.0
whitenoise==6.7.0
gunicorn==23.0.0

# OpenTelemetry
opentelemetry-distro==0.45b0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-django==0.45b0
# opentelemetry-instrumentation-djangorestframework # Not supported on Python 3.13 yet
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-psycopg2==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-sdk==1.24.0

# Metrics
django-prometheus==2.3.1

# Logging
python-json-logger==2.0.7

# Metrics
django-prometheus==2.3.1